refactor: change app links component location
