+++ app/routes/_index.tsx
@@ -3,12 +3,12 @@ import { useLoaderData, useNavigation, useSubmit } from '@remix-run/react';
 import { hc } from 'hono/client';
 import { useEffect, useState } from 'react';
 import { AppType } from 'server';
-import { AppGrid } from '~/feature/app-links/components/app-grid';
-import { EmptyState } from '~/feature/app-links/components/empty-state';
-import { ErrorState } from '~/feature/app-links/components/error-state';
-import { Header } from '~/feature/app-links/components/header';
-import { LoadingState } from '~/feature/app-links/components/loading-state';
-import { Pagination } from '~/feature/app-links/components/pagination';
+import { AppGrid } from '~/components/app-links/app-grid';
+import { EmptyState } from '~/components/app-links/empty-state';
+import { ErrorState } from '~/components/app-links/error-state';
+import { Header } from '~/components/app-links/header';
+import { LoadingState } from '~/components/app-links/loading-state';
+import { Pagination } from '~/components/app-links/pagination';
 import { usePagination } from '~/hooks/use-pagination';
 
 const client = hc<AppType>(import.meta.env.VITE_API_URL);
