+++ app/routes/admin/index.tsx
@@ -4,11 +4,11 @@ import { hc } from 'hono/client';
 import { RefreshCw } from 'lucide-react';
 import { useEffect, useState } from 'react';
 import { AppType } from 'server';
+import { Pagination } from '~/components/app-links/pagination';
 import { Button } from '~/components/ui/button';
 import { Checkbox } from '~/components/ui/checkbox';
 import { SearchInput } from '~/components/ui/search-input';
 import { Spinner } from '~/components/ui/spinner';
-import { Pagination } from '~/feature/app-links/components/pagination';
 import { usePagination } from '~/hooks/use-pagination';
 import { useToast } from '~/hooks/use-toast';
 import { AddLinkButton } from '../add-link';
